Renamed bioc module to distinguish the standard as AC-bioc
